add process manager persistance file, ensure loaded
