+++ lib/commanded/storage/eventstore/connection.ex
@@ -1,0 +1,6 @@
+if Code.ensure_loaded?(Eventstore) do
+
+  defmodule Commanded.Storage.Eventstore.Connection do
+  end
+
+end
