+++ lib/commanded/storage/postgre/connection.ex
@@ -1,0 +1,5 @@
+if Code.ensure_loaded?(Postgre) do
+
+  defmodule Commanded.Storage.Postgre.Connection do
+  end
+end
