+++ test/aggregates/persistence_test.exs
@@ -1,8 +1,5 @@
-defmodule Commanded.Storage.TestDbs do
+defmodule Commanded.Aggregates.PersistenceTest do
   use Commanded.StorageCase
-
-  import Commanded.Enumerable, only: [pluck: 2]
-  alias Commanded.Aggregates.{Registry,Aggregate}
 
   defmodule ExampleAggregate do
     defstruct [
@@ -17,8 +14,15 @@ defmodule Commanded.Storage.TestDbs do
     alias Commands.{AppendItems}
     alias Events.{ItemAppended}
 
-    def append_items(%ExampleAggregate{last_index: last_index}, count), do:
-      Enum.map(1..count, fn index -> %ItemAppended{index: last_index + index} end)
+    def append_items(%ExampleAggregate{last_index: last_index}, count) do
+      Enum.map(1..count, fn index ->
+        %ItemAppended{index: last_index + index}
+      end)
+    end
+
+    def append_item(%ExampleAggregate{last_index: last_index}, %AppendItems{count: count}) do
+      %ItemAppended{index: last_index + 1}
+    end
 
     # state mutatators
     def apply(%ExampleAggregate{items: items} = state, %ItemAppended{index: index}) do
@@ -30,17 +34,21 @@ defmodule Commanded.Storage.TestDbs do
   end
 
   alias ExampleAggregate.Commands.{AppendItems}
-
-  # handler
-  defmodule AppendItemsHandler do
-    @behaviour Commanded.Commands.Handler
-    def handle(%ExampleAggregate{} = aggregate, %AppendItems{count: count}), do:
-      ExampleAggregate.append_items(aggregate, count)
-  end
+  alias Commanded.Aggregates.Persistence
 
 
-  test "should be true" do
-    assert 1 == 1
+
+  test "Apply events for a data structure" do
+
+    aggregate = %ExampleAggregate{}
+
+    events = ExampleAggregate.append_items(aggregate, 5)
+
+    res = Commanded.Aggregates.Persistence.apply_events(ExampleAggregate, aggregate, events)
+
+
+    last_state = %ExampleAggregate{items: [1, 2, 3, 4, 5], last_index: 5}
+    assert res = last_state
   end
   # test "should persist pending events in order applied" do
   #   aggregate_uuid = UUID.uuid4
